Refactor and relocate column name fix tool to TreeOfLife toolbox
Moved the `column_name_change_lila_fix` tool from `DD_tools` to `TreeOfLife_toolbox`, updating all imports, configurations, and documentation accordingly. Enhanced class docstrings and added comprehensive explanations to improve code clarity and maintainability.
