--- src/DD_tools/column_name_change_lila_fix/README.md
@@ -1,5 +1,0 @@
-tool to fix the column name mistake in `lila-bc-multi-label`
-
-Additional config fields:
-
-* `uuid_table_path` - path to uuid table with duplicated uuids
